+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Browser:
 
 - secure/serverlist.html - Added input element and checkbox element for new properties "altNick" and "recoverNick"
 - secure/js/serverlist.js - Added value parsing for new server properties "altNick" and "recoverNick"
+- secure/js/webclient.js - On change in ircState object, nickname input element updates from state object even if not connected to IRC. This is necessary for alternate nickname rotation and nickname recovery. However, a change in state could possibly revert the nickname input element to current value on the web server. 
 
 ## [v0.2.4](https://github.com/cotarr/irc-hybrid-client/releases/tag/v0.2.4) 2022-08-04